rename tests to comply with nosetests
nosetest tests discovery won't recognize tests in files named
_test_object.py

also, nosetests with run a test twice if it imported somewhere else,
e.g. if it was imported in __init__.py.
