+++ pele/accept_tests/tests/__init__.py
@@ -1,3 +1,1 @@
-from dont_leave_basin_test import *
-from spherical_container_test import *
-from metropolis_test import *
+
