+++ pele/mindist/tests/__init__.py
@@ -1,3 +1,1 @@
-from minpermdist_stochastic_test import *
-from permutational_alignment_test import *
-from _test_exact_match_cluster import *
+
