+++ pele/optimize/tests/__init__.py
@@ -1,11 +1,1 @@
-from _test_wrappers import *
-from _test_with_functions import *
-from _test_lbfgs import *
-from _test_mylbfgs import *
-from _test_mylbfgs_vs_lbfgs import *
-from _test_lbfgs_cpp import *
-from _test_modified_fire_cpp import *
 
-
-if __name__ == "__main__":
-    unittest.main()
