+++ pele/optimize/tests/test_lbfgs_cpp.py
