+++ pele/optimize/tests/test_modified_fire_cpp.py
