+++ pele/optimize/tests/test_with_functions.py
