+++ pele/optimize/tests/test_wrappers.py
