+++ pele/potentials/tests/__init__.py
@@ -1,15 +1,1 @@
-from _test_beale import *
-from _test_booth import *
-from _test_lj_cpp import *
-from _test_wca_cpp import *
-from _test_hs_wca_cpp import *
-from _test_python_cpp_wrapper import *
-from _test_blj_cpp import *
-from _test_morse_cpp import *
-from _test_lj import *
-from _test_ljcut import *
-from _test_atlj import *
 
-
-if __name__ == "__main__":
-    unittest.main()
