+++ pele/potentials/tests/test_beale.py
