+++ pele/potentials/tests/test_booth.py
