+++ pele/potentials/tests/test_lj.py
