+++ pele/potentials/tests/test_lj_cpp.py
