+++ pele/potentials/tests/test_morse_cpp.py
