+++ pele/rates/tests/__init__.py
@@ -1,7 +1,1 @@
-from test_graph_transformation import *
-from test_graph_transformation_minima import *
-from test_kmc import *
-from test_cpp_ngt import * 
 
-if __name__ == "__main__":
-    unittest.main()
