+++ pele/storage/tests/__init__.py
@@ -1,5 +1,1 @@
-from test_database import *
-# from test_migrate_db import *
 
-if __name__ == "__main__":
-    unittest.main()
