+++ pele/systems/tests/__init__.py
@@ -1,6 +1,1 @@
-from _test_frozen_lj import *
-from _test_ljcluster import *
-from _test_bljcluster import *
 
-if __name__ == "__main__":
-    unittest.main()
