+++ pele/transition_states/tests/__init__.py
@@ -1,10 +1,1 @@
-from _test_eig_pot import *
-from _test_find_lowest_eigenvector import *
-from _test_generalized_dimer import *
-from _test_find_transition_state_hef import *
-from _test_transverse_walker import *
-from _test_dimer_translator import *
-#from _test_generalized_hef import *
 
-if __name__ == "__main__":
-    unittest.main()
