+++ pele/transition_states/tests/test_eig_pot.py
