+++ pele/transition_states/tests/test_find_transition_state_hef.py
@@ -45,7 +45,7 @@ class TestHEF_InvertedGradient(TestFindTransitionState):
 
 class TestFindTransitionState_NFEV(unittest.TestCase):
     def setUp(self):
-        from pele.optimize.tests._test_nfev import _PotWrapper
+        from pele.optimize.tests.test_nfev import _PotWrapper
         self.system = LJCluster(18)
         self.pot = _PotWrapper(self.system.get_potential())
         self.x = self.system.get_random_minimized_configuration(tol=10.).coords
