+++ pele/transition_states/tests/test_generalized_dimer.py
