+++ pele/transition_states/tests/test_generalized_hef.py
@@ -5,10 +5,10 @@ import nose
 from pele.systems import LJCluster
 from pele.transition_states import GeneralizedDimer
 from pele.utils.xyz import read_xyz
-import _test_generalized_dimer
+import test_generalized_dimer
 
-nose.tools.nottest
-class TestGeneralizedHEF(_test_generalized_dimer.TestGeneralizedDimer):
+@nose.tools.nottest
+class TestGeneralizedHEF(test_generalized_dimer.TestGeneralizedDimer):
     
     def make_dimer(self, x):
         return GeneralizedDimer(x, self.pot,
