+++ pele/transition_states/tests/test_transverse_walker.py
@@ -28,7 +28,7 @@ class TestTransverseWalker(unittest.TestCase):
 
 class TestTransverseWalker_NFEV(unittest.TestCase):
     def setUp(self):
-        from pele.optimize.tests._test_nfev import _PotWrapper
+        from pele.optimize.tests.test_nfev import _PotWrapper
         self.system = LJCluster(18)
         self.pot = _PotWrapper(self.system.get_potential())
