+++ pele/utils/tests/__init__.py
@@ -1,6 +1,1 @@
-from test_disconnectivity_graph import *
-from test_frozen_potential import *
 
-
-if __name__ == "__main__":
-    unittest.main()
